fix(web): lazy-load admin-analytics so Recharts stays out of initial bundle (#187)
Bundle size check (rule B11) failed on the previous commit: the
admin-analytics route was statically wired through routeTree.gen.ts,
pulling Recharts (~110 kB gzip) into the initial chunk and pushing
the payload to 636 kB / 570 kB budget.

Split the page into the standard TanStack Router lazy pattern: a
minimal index.tsx with createFileRoute({}), and an index.lazy.tsx
with createLazyFileRoute holding the component. Recharts now ships
in generateCategoricalChart-*.js (102.9 kB) loaded on demand when an
admin actually visits /admin-analytics.

Initial bundle: 560.6 kB / 570 kB (was 636.1 kB).
